Refactor streamlit app
